requirements.txt
matplotlib==3.4.3
matplotlib-inline==0.1.6
numpy==1.21.2
openpyxl==3.1.2
pandas==1.3.2
Pillow==8.2.0
plotly==5.14.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.1
sklearn==0.0
streamlit==1.22.0
streamlit-ext==0.1.7
yellowbrick==1.5

